Make UserInfo outer class
